+++ src/Plugin/Loadbalancer/Event/EndpointFailure.php
@@ -4,7 +4,7 @@ namespace Solarium\Plugin\Loadbalancer\Event;
 
 use Solarium\Core\Client\Endpoint;
 use Solarium\Exception\HttpException;
-use Symfony\Component\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * EndpointFailure event, see Events for details.